+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/25)
 * Add carbon tracker icon
+* Add header with traceId (uuid) to each api call
 
 0.34 (2024-10-07)
 ===============